[FE][refactor] : 마일스톤 컴포넌트 분리 #21
마일스톤 컴포넌트 분리
버튼 눌릴 때 style 변화
