+++ frontend/pages/milestones/MilestoneListPage.jsx
@@ -6,6 +6,7 @@ const MilestoneListPage = () => {
   const [open, setOpenMilestone] = useState([]);
   const [close, setClosedMilestone] = useState([]);
   const [milestoneList, setMilestoneList] = useState([]);
+  const [state, setState] = useState('open');
   const TRUE = 1;
   const FALSE = 0;
   const getOpen = async () => {
@@ -22,6 +23,7 @@ const MilestoneListPage = () => {
         ? await service.getMilestones({ isClosed: TRUE })
         : await service.getMilestones({ isClosed: FALSE });
     setMilestoneList(milestones.data);
+    setState(status);
   };
   useEffect(() => {
     getOpen();
@@ -30,7 +32,13 @@ const MilestoneListPage = () => {
   }, []);
   return (
     <>
-      <MilestoneList milestoneList={milestoneList} open={open} close={close} getMilestones={getMilestones} />
+      <MilestoneList
+        milestoneList={milestoneList}
+        state={state}
+        open={open}
+        close={close}
+        getMilestones={getMilestones}
+      />
     </>
   );
 };
